Improve acceptance tests for route table import
Use checks that align with inline routes in the state. Improve
tests to avoid network values that will conflict with fewer AWS
accounts.

Previously, route table import acceptance tests only checked the
length of the state (1 for route table, and 1 for each separate
route resource). This simple test did not check much, especially
when checking route tables with inline routes.

Now, the state is checked to make sure it has routes and a route-
count attribute that matches to expected value.